Updated Destroy Stage

+++ Jenkinsfile
@@ -57,15 +57,29 @@ pipeline {
             }
         }
 
-        stage('Destroy') {
-          when { anyOf
-                        {
-                          environment name: 'ACTION', value: 'destroy'
-                        }
-          }
-          steps {
-                sh "terraform destroy -input=false --var-file=env/${params.environment}.tfvars -auto-approve"
-            }
+        stage('Destroy') {    
+    			when { anyOf
+    					{
+    						environment name: 'ACTION', value: 'destroy';
+    					}
+    				}
+    			steps {
+    				script {
+    					def IS_APPROVED = input(
+    						message: "Destroy ${ENV_NAME} !?!",
+    						ok: "Yes",
+    						parameters: [
+    							string(name: 'IS_APPROVED', defaultValue: 'No', description: 'Think again!!!')
+    						]
+    					)
+    					if (IS_APPROVED != 'Yes') {
+    						currentBuild.result = "ABORTED"
+    						error "User cancelled"
+    					} else {
+    						sh "terraform destroy -input=false --var-file=env/${params.environment}.tfvars -auto-approve"
+    				}
+    			}
         }
     }
+}
 }
